requirements.txt
langchain==0.3.1
langchain_community==0.3.1
langchain-openai==0.2.0
transformers==4.46.3
bitsandbytes==0.44.1
minio==7.2.12
pdfplumber==0.11.4
python-pptx==1.0.2
unstructured==0.16.4
python-docx==1.1.2
openpyxl==3.1.5
pypdf==4.3.1
pandas==2.2.3
olefile==0.47
tabulate==0.9.0